backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt
django-cors-headers==4.3.1
drf-yasg==1.21.7
setuptools
django-filter